vq-vaePractice/requirements.txt
python==3.10
pytorch==2.1.2
torchvision==0.17.1
torchaudio==2.2.1
cudatoolkit=11.8